+++ src/pages/registrationManagementPage/styles/RegistrationManagementPage.styles.js
@@ -1,0 +1,13 @@
+const { default: styled } = require("styled-components");
+
+export const Wrap = styled.div`
+  display: flex;
+  height: 100vh;
+`;
+
+export const ViewBox = styled.div`
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  width: 100%;
+`;
