+++ src/types/EnumRegistrationManagementScreen.js
@@ -1,0 +1,8 @@
+export const EnumRegistrationManagementScreen = {
+    ScreenTeamRegister: 'ScreenTeamRegister',
+    ScreenPlayerRegister: 'ScreenPlayerRegister',
+    ScreenCodeManagement: 'ScreenCodeManagement',
+    ScreenRefereeRegister: 'ScreenRefereeRegister',
+    ScreenGameSelect: 'ScreenGameSelect',
+    ScreenGameInformManagement: 'ScreenGameInformManagement',
+}
